+++ source/parser/core/interpolator.js
@@ -1,20 +1,20 @@
-import op from '../lib/object-path.js'
-import re from "../lib/re.js";
+import op from '#lib/object-path.js'
+import re from "#lib/re.js";
 
 
 export default function (config)
 {
-    let { interpolation, vars } = config;
+    let { interpolation } = config;
 
     let varis = re.g(interpolation.variables);
     let sects = re.g(interpolation.segments);
     let check = (one, two) => one === two ? '' : two
     
-    return (str, data) =>
+    let inter = (str, data) =>
     {
         str ??= '';
 
-        let reps = { ...vars, ...data };
+        let reps = { ...config.vars, ...data };
         let solve = value => typeof value === 'function' ? value(data) : value
 
         let doVars = str => str.replace(varis, (match, name, def) => solve(op.get(reps, name)) ?? def ?? match)
@@ -24,4 +24,21 @@ export default function (config)
 
         return doVars(str);
     }
+
+    /*
+        Creates an interpolation function from a `spec`.
+
+        The spec can be a string or function.
+    */
+    inter.toFunc = spec =>
+    {
+        // spec function is called with data and vars
+        if (typeof spec === 'function') return data => inter(spec(data, config.vars), data);        
+        // spec string is directly interpolated
+        if (typeof spec === 'string') return data => inter(spec, data)        
+        // no output for invalid spec
+        return () => ''
+    }
+
+    return inter;
 }